requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
imageio==2.34.1
iniconfig==2.0.0
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
tifffile==2024.5.22
